+++ cfgrib/cfmessage.py
@@ -18,9 +18,6 @@
 #   Alessandro Amici - B-Open - https://bopen.eu
 #
 
-from __future__ import absolute_import, division, print_function, unicode_literals
-from builtins import int, str
-
 import datetime
 import logging
 import typing as T  # noqa