+++ cfgrib/dataset.py
@@ -17,9 +17,6 @@
 #   Alessandro Amici - B-Open - https://bopen.eu
 #
 
-from __future__ import absolute_import, division, print_function, unicode_literals
-from builtins import float, int, list, object, set, str
-
 import collections
 import datetime
 import json
@@ -460,11 +457,10 @@ def build_dataset_components(
             raise DatasetBuildError(error_message, key, fbks)
         if 'parameter' in encode_cf and var_name not in ('undef', 'unknown'):
             short_name = var_name
-        vars = collections.OrderedDict([(short_name, data_var)])
-        vars.update(coord_vars)
         try:
+            dict_merge(variables, coord_vars)
+            dict_merge(variables, {short_name: data_var})
             dict_merge(dimensions, dims)
-            dict_merge(variables, vars)
         except ValueError:
             if errors == 'ignore':
                 pass