+++ cfgrib/xarray_to_grib.py
@@ -19,8 +19,6 @@
 #   Leonardo Barcaroli - B-Open - https://bopen.eu
 #
 
-from __future__ import absolute_import, division, print_function, unicode_literals
-
 import itertools
 import logging
 import typing as T  # noqa